+++ src/common/Widgets/GRadioGroup.cpp
@@ -683,7 +683,7 @@ bool GRadioButton::OnKey(GKey &k)
 			}
 			if (Btns.Length() > 1)
 			{
-				int Index = Btns.IndexOf(this);
+				ssize_t Index = Btns.IndexOf(this);
 				if (Index >= 0)
 				{
 					GRadioButton *n = Btns[(Index + Move + Btns.Length()) % Btns.Length()];
